+++ zebra-rpc/src/methods/get_block_template_rpcs.rs
@@ -38,7 +38,9 @@ use crate::methods::{
 
 pub mod config;
 pub mod constants;
+
 pub(crate) mod types;
+pub(crate) mod zip317;
 
 /// The max estimated distance to the chain tip for the getblocktemplate method
 // Set to 30 in case the local time is a little ahead.
@@ -310,7 +312,7 @@ where
                 });
             }
 
-            let mempool_txs = select_mempool_transactions(mempool).await?;
+            let mempool_txs = zip317::select_mempool_transactions(mempool).await?;
 
             let miner_fee = miner_fee(&mempool_txs);
 
@@ -447,37 +449,6 @@ where
 }
 
 // get_block_template support methods
-
-/// Returns selected transactions in the `mempool`, or an error if the mempool has failed.
-///
-/// TODO: select transactions according to ZIP-317 (#5473)
-pub async fn select_mempool_transactions<Mempool>(
-    mempool: Mempool,
-) -> Result<Vec<VerifiedUnminedTx>>
-where
-    Mempool: Service<
-            mempool::Request,
-            Response = mempool::Response,
-            Error = zebra_node_services::BoxError,
-        > + 'static,
-    Mempool::Future: Send,
-{
-    let response = mempool
-        .oneshot(mempool::Request::FullTransactions)
-        .await
-        .map_err(|error| Error {
-            code: ErrorCode::ServerError(0),
-            message: error.to_string(),
-            data: None,
-        })?;
-
-    if let mempool::Response::FullTransactions(transactions) = response {
-        // TODO: select transactions according to ZIP-317 (#5473)
-        Ok(transactions)
-    } else {
-        unreachable!("unmatched response to a mempool::FullTransactions request");
-    }
-}
 
 /// Returns the total miner fee for `mempool_txs`.
 pub fn miner_fee(mempool_txs: &[VerifiedUnminedTx]) -> Amount<NonNegative> {
